+++ background.js
@@ -1,5 +1,5 @@
 // browserAction behaviour
-browser.browserAction.setTitle({ title:  'Open tab events log' });
+browser.browserAction.setTitle({ title: 'Open tab events log' });
 browser.browserAction.onClicked.addListener(() => {
   browser.tabs.create({ url: 'page.html' });
 });