+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## v0.1.2 – 2020-03-06 (Unreleased)
+## v0.1.2 – 2020-03-07
 
-Add entry groups
+* Add options for favicon, event grouping and window focus events.
+* Bump Firefox version for Javascript features.
+* Add entry groups
 
 ## v0.1.1 – 2020-03-06